Add navbar for landing page
