+++ package.json
@@ -2,7 +2,7 @@
   "name": "degreely",
   "version": "0.1.0",
   "private": true,
-  "homepage": "http://hellodommy.github.io/degreely",
+  "homepage": "https://degreely.github.io/degreely/",
   "dependencies": {
     "-": "0.0.1",
     "@material-ui/core": "^4.11.0",
@@ -11,6 +11,7 @@
     "@testing-library/user-event": "^7.1.2",
     "bootstrap": "^4.5.3",
     "gh-pages": "^3.1.0",
+    "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-beautiful-dnd": "^13.0.0",
     "react-bootstrap": "^1.3.0",
