--- src/components/LandingHeader.js
@@ -1,34 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import Navbar from "react-bootstrap/Navbar";
-import Nav from "react-bootstrap/Nav";
-import logo from "../img/logo.svg";
-import "../css/LandingHeader.css";
-
-function LandingHeader() {
-  return (
-    <div className="LandingHeader">
-      <Navbar bg="light" expand="lg">
-        <Navbar.Brand href="#home">
-          <img src={logo} className="App-logo" alt="logo" />
-        </Navbar.Brand>
-        <Navbar.Toggle aria-controls="basic-navbar-nav" />
-        <Navbar.Collapse id="basic-navbar-nav">
-          <Nav className="mr-auto">
-            <Nav.Link as={Link} to="/">
-              Home
-            </Nav.Link>
-            <Nav.Link as={Link} to="/about">
-              About
-            </Nav.Link>
-            <Nav.Link as={Link} to="/support">
-              Support
-            </Nav.Link>
-          </Nav>
-        </Navbar.Collapse>
-      </Navbar>
-    </div>
-  );
-}
-
-export default LandingHeader;
