--- src/css/LandingHeader.css
@@ -1,12 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 32px;
-  pointer-events: none;
-}
-
-.App-link {
-  color: #61dafb;
-}
