+++ src/scss/App.scss
@@ -1,0 +1,3 @@
+body {
+    background-color: #F5F8FF;
+}
