+++ src/scss/LandingHeader.scss
@@ -1,0 +1,12 @@
+.App-logo {
+  height: 32px;
+  pointer-events: none;
+}
+
+.Nav-link {
+    color: #333;
+}
+
+.Landing-header {
+  background-color: #F5F8FF;
+}
